backend/requirements.txt
Django==6.0.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient>=2.2.1
